Add CreateForm; Move forms to pages directory
